+++ src/net/minecraft/src/PacketHooksBootstrapMcPacketSniffer.java
@@ -1,0 +1,10 @@
+package net.minecraft.src;
+
+import com.bencvt.minecraft.mcpacketsniffer.Controller;
+
+/** This class is automagically instantiated by PacketHooks. */
+public class PacketHooksBootstrapMcPacketSniffer {
+    public PacketHooksBootstrapMcPacketSniffer() {
+        Controller.getInstance();
+    }
+}