+++ src/net/minecraft/src/mod_McPacketSniffer.java
@@ -2,10 +2,15 @@ package net.minecraft.src;
 
 import com.bencvt.minecraft.mcpacketsniffer.Controller;
 
+/**
+ * ModLoader mod definition, not really necessary because we're already
+ * creating the Controller in {@link PacketHooksBootstrapMcPacketSniffer}.
+ * But keep this definition around so it shows up in the user's list of mods.
+ */
 public class mod_McPacketSniffer extends BaseMod {
     @Override
     public void load() {
-        Controller.getInstance();
+        Controller.getInstance(); // redundant but harmless
     }
 
     @Override